Move the lozad observer outside the artistitem component
